+++ ton_client/src/json_interface/modules.rs
@@ -257,6 +257,10 @@ fn register_abi(handlers: &mut RuntimeHandlers) {
         crate::abi::encode_message::encode_message_api,
     );
     module.register_async_fn(
+        crate::abi::encode_internal_message,
+        crate::abi::encode_message::encode_internal_message_api,
+    );
+    module.register_async_fn(
         crate::abi::attach_signature,
         crate::abi::encode_message::attach_signature_api,
     );