+++ ton_sdk/src/contract.rs
@@ -21,10 +21,7 @@ use serde_json::Value;
 use std::convert::Into;
 use std::io::{Cursor, Read, Seek};
 use ton_abi::json_abi::DecodedMessage;
-use ton_block::{
-    AccountIdPrefixFull, Deserializable, ExternalInboundMessageHeader, GetRepresentationHash,
-    Message as TvmMessage, MsgAddressInt, Serializable, ShardIdent, StateInit,
-};
+use ton_block::{AccountIdPrefixFull, Deserializable, ExternalInboundMessageHeader, GetRepresentationHash, Message as TvmMessage, MsgAddressInt, Serializable, ShardIdent, StateInit, InternalMessageHeader, CurrencyCollection};
 use ton_types::cells_serialization::deserialize_cells_tree;
 use ton_types::{error, AccountId, Result, SliceData};
 
@@ -299,13 +296,12 @@ impl Contract {
 
     // ------- Call constructing functions -------
 
-    // Packs given inputs by abi into Message struct.
+    // Packs given inputs by abi into an external inbound Message struct.
     // Works with json representation of input and abi.
     // Returns message's bag of cells and identifier.
-    pub fn construct_call_message_json(
+    pub fn construct_call_ext_in_message_json(
         address: MsgAddressInt,
         params: FunctionCallSet,
-        internal: bool,
         key_pair: Option<&Keypair>,
     ) -> Result<SdkMessage> {
         // pack params into bag of cells via ABI
@@ -314,11 +310,41 @@ impl Contract {
             params.func,
             params.header,
             params.input,
-            internal,
+            false,
             key_pair,
         )?;
 
-        let msg = Self::create_message(address.clone(), msg_body.into())?;
+        let msg = Self::create_ext_in_message(address.clone(), msg_body.into())?;
+        let (body, id) = Self::serialize_message(&msg)?;
+        Ok(SdkMessage {
+            id,
+            serialized_message: body,
+            message: msg,
+            address,
+        })
+    }
+
+    // Packs given inputs by abi into an internal Message struct.
+    // Works with json representation of input and abi.
+    // Returns message's bag of cells and identifier.
+    pub fn construct_call_int_message_json(
+        address: MsgAddressInt,
+        ihr_disabled: bool,
+        bounce: bool,
+        value: CurrencyCollection,
+        params: FunctionCallSet,
+    ) -> Result<SdkMessage> {
+        // pack params into bag of cells via ABI
+        let msg_body = ton_abi::encode_function_call(
+            params.abi,
+            params.func,
+            None,
+            params.input,
+            true,
+            None,
+        )?;
+
+        let msg = Self::create_int_message(ihr_disabled, bounce, address.clone(), value, msg_body.into())?;
         let (body, id) = Self::serialize_message(&msg)?;
         Ok(SdkMessage {
             id,
@@ -343,7 +369,7 @@ impl Contract {
             params.input,
         )?;
 
-        let msg = Self::create_message(address, msg_body.into())?;
+        let msg = Self::create_ext_in_message(address, msg_body.into())?;
 
         Self::serialize_message(&msg).map(|(msg_data, _id)| MessageToSign {
             message: msg_data,
@@ -372,7 +398,7 @@ impl Contract {
         )?;
 
         let cell = msg_body.into();
-        let msg = Self::create_deploy_message(Some(cell), image, workchain_id)?;
+        let msg = Self::create_ext_deploy_message(Some(cell), image, workchain_id)?;
 
         let address = msg.dst().ok_or_else(|| {
             error!(SdkError::InternalError {
@@ -401,20 +427,31 @@ impl Contract {
             Some(data) => Some(Self::deserialize_tree_to_slice(data)?),
         };
 
-        Self::create_deploy_message(body_cell, image, workchain_id)
+        Self::create_ext_deploy_message(body_cell, image, workchain_id)
     }
 
-    // Packs given image into Message struct.
+    // Packs given image into an external inbound Message struct.
     // Returns message's bag of cells and identifier.
     pub fn construct_deploy_message_no_constructor(
         image: ContractImage,
         workchain_id: i32,
     ) -> Result<TvmMessage> {
-        Self::create_deploy_message(None, image, workchain_id)
+        Self::create_ext_deploy_message(None, image, workchain_id)
     }
 
-    // Packs given image and input into Message struct without sign and returns data to sign.
-    // Sign should be then added with `add_sign_to_message` function
+    // Packs given image into an internal Message struct.
+    // Returns message's bag of cells and identifier.
+    pub fn construct_int_deploy_message_no_constructor(
+        image: ContractImage,
+        workchain_id: i32,
+        ihr_disabled: bool,
+        bounce: bool,
+    ) -> Result<TvmMessage> {
+        Self::create_int_deploy_message(None, image, workchain_id, ihr_disabled, bounce)
+    }
+
+    // Packs given image and input into Message struct without signature and returns data to sign.
+    // Signature should be then added with `add_sign_to_message` function
     // Works with json representation of input and abi.
     pub fn get_deploy_message_bytes_for_signing(
         params: FunctionCallSet,
@@ -429,12 +466,37 @@ impl Contract {
         )?;
 
         let cell = msg_body.into();
-        let msg = Self::create_deploy_message(Some(cell), image, workchain_id)?;
+        let msg = Self::create_ext_deploy_message(Some(cell), image, workchain_id)?;
 
         Self::serialize_message(&msg).map(|(msg_data, _id)| MessageToSign {
             message: msg_data,
             data_to_sign,
         })
+    }
+
+    // Packs given image and input into Message struct with internal header and returns data.
+    // Works with json representation of input and abi.
+    pub fn get_int_deploy_message_bytes(
+        params: FunctionCallSet,
+        image: ContractImage,
+        workchain_id: i32,
+        ihr_disabled: bool,
+        bounce: bool,
+    ) -> Result<Vec<u8>> {
+        let msg_body = ton_abi::encode_function_call(
+            params.abi,
+            params.func,
+            None,
+            params.input,
+            true,
+            None,
+        )?;
+
+        let cell = msg_body.into();
+        let msg = Self::create_int_deploy_message(Some(cell), image, workchain_id, ihr_disabled, bounce)?;
+
+        Self::serialize_message(&msg)
+            .map(|(msg_data, _id)| msg_data)
     }
 
     // Add sign to message, returned by `get_deploy_message_bytes_for_signing` or
@@ -507,7 +569,7 @@ impl Contract {
         })
     }
 
-    fn create_message(address: MsgAddressInt, msg_body: SliceData) -> Result<TvmMessage> {
+    fn create_ext_in_message(address: MsgAddressInt, msg_body: SliceData) -> Result<TvmMessage> {
         let mut msg_header = ExternalInboundMessageHeader::default();
         msg_header.dst = address;
 
@@ -517,16 +579,62 @@ impl Contract {
         Ok(msg)
     }
 
-    fn create_deploy_message(
+    fn create_int_message(
+        ihr_disabled: bool,
+        bounce: bool,
+        dst: MsgAddressInt,
+        value: CurrencyCollection,
+        msg_body: SliceData,
+    ) -> Result<TvmMessage> {
+        let msg_header = InternalMessageHeader {
+            ihr_disabled,
+            bounce,
+            dst,
+            value,
+            ..Default::default()
+        };
+
+        let mut msg = TvmMessage::with_int_header(msg_header);
+        msg.set_body(msg_body);
+
+        Ok(msg)
+    }
+
+    pub(crate) fn create_ext_deploy_message(
         msg_body: Option<SliceData>,
         image: ContractImage,
         workchain_id: i32,
     ) -> Result<TvmMessage> {
-        let mut msg_header = ExternalInboundMessageHeader::default();
-        msg_header.dst = image.msg_address(workchain_id);
+        let msg_header = ExternalInboundMessageHeader {
+            dst: image.msg_address(workchain_id),
+            ..Default::default()
+        };
+
         let mut msg = TvmMessage::with_ext_in_header(msg_header);
         msg.set_state_init(image.state_init());
         msg_body.map(|body| msg.set_body(body));
+
+        Ok(msg)
+    }
+
+    pub(crate) fn create_int_deploy_message(
+        msg_body: Option<SliceData>,
+        image: ContractImage,
+        workchain_id: i32,
+        ihr_disabled: bool,
+        bounce: bool,
+    ) -> Result<TvmMessage> {
+        let msg_header = InternalMessageHeader {
+            ihr_disabled,
+            bounce,
+            dst: image.msg_address(workchain_id),
+            ..Default::default()
+        };
+
+        let mut msg = TvmMessage::with_int_header(msg_header);
+        msg.set_state_init(image.state_init());
+        msg_body.map(|body| msg.set_body(body));
+
         Ok(msg)
     }