+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 and decerializations which drastically improves performance of `run_tvm` and `run_executor` expecially in case of numerous calls on the same data. 
 - `boc_cache` parameter in `tvm.run_tvm` and `tvm.run_executor` functions to save resulting messages and account BOCs into cache.
 - `return_updated_account` flag parameter introduced in `tvm.run_tvm` and `tvm.run_executor` functions to return updated account state. Important: by default this flag is `false` and account data is not returned.
+- `abi.encode_internal_message` function to encode an internal ABI-compatible message.
 
 ## 1.6.3 Feb 4, 2021
 ### Fixed